+++ packages/calcite-components/src/components/date-picker-day/date-picker-day.e2e.ts
@@ -1,4 +1,5 @@
-import { E2EPage } from "@stencil/core/testing";
+import { E2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
+import { describe, expect, it, beforeEach } from "vitest";
 import { disabled } from "../../tests/commonTests";
 import { newProgrammaticE2EPage } from "../../tests/utils";
 import { DATE_PICKER_FORMAT_OPTIONS } from "../date-picker/resources";