+++ packages/calcite-components/src/components/split-button/split-button.tsx
@@ -1,14 +1,5 @@
-import {
-  Component,
-  Element,
-  Event,
-  EventEmitter,
-  h,
-  Method,
-  Prop,
-  VNode,
-  Watch,
-} from "@stencil/core";
+import { PropertyValues } from "lit";
+import { LitElement, property, createEvent, h, method, JsxNode } from "@arcgis/lumina";
 import { FlipPlacement, MenuPlacement, OverlayPositioning } from "../../utils/floating-ui";
 import {
   InteractiveComponent,
@@ -24,74 +15,73 @@ import {
 import { DropdownIconType } from "../button/interfaces";
 import { Appearance, FlipContext, Kind, Scale, Width } from "../interfaces";
 import { IconNameOrString } from "../icon/interfaces";
+import { focusFirstTabbable } from "../../utils/dom";
 import { CSS } from "./resources";
+import { styles } from "./split-button.scss";
 
-/**
- * @slot - A slot for adding `calcite-dropdown` content.
- */
-@Component({
-  tag: "calcite-split-button",
-  styleUrl: "split-button.scss",
-  shadow: {
-    delegatesFocus: true,
-  },
-})
-export class SplitButton implements InteractiveComponent, LoadableComponent {
-  //--------------------------------------------------------------------------
-  //
-  //  Public Properties
-  //
-  //--------------------------------------------------------------------------
-
-  /** Specifies the appearance style of the component. */
-  @Prop({ reflect: true }) appearance: Extract<
-    "outline" | "outline-fill" | "solid" | "transparent",
-    Appearance
-  > = "solid";
-
-  /** Specifies the kind of the component, which will apply to border and background, if applicable. */
-  @Prop({ reflect: true }) kind: Extract<"brand" | "danger" | "inverse" | "neutral", Kind> =
-    "brand";
-
-  /** When `true`, interaction is prevented and the component is displayed with lower opacity. */
-  @Prop({ reflect: true }) disabled = false;
-
-  @Watch("disabled")
-  handleDisabledChange(value: boolean): void {
-    if (!value) {
-      this.active = false;
-    }
+declare global {
+  interface DeclareElements {
+    "calcite-split-button": SplitButton;
   }
+}
+
+/** @slot - A slot for adding `calcite-dropdown` content. */
+export class SplitButton extends LitElement implements InteractiveComponent, LoadableComponent {
+  // #region Static Members
+
+  static override shadowRootOptions = { mode: "open" as const, delegatesFocus: true };
+
+  static override styles = styles;
+
+  // #endregion
+
+  // #region Private Properties
+
+  private get dropdownIcon(): string {
+    return this.dropdownIconType === "chevron"
+      ? "chevronDown"
+      : this.dropdownIconType === "caret"
+        ? "caretDown"
+        : this.dropdownIconType === "ellipsis"
+          ? "ellipsis"
+          : "handle-vertical";
+  }
+
+  // #endregion
+
+  // #region Public Properties
 
   /**
    * When `true`, the component is active.
    *
-   * @internal
+   * @private
    */
-  @Prop({ mutable: true, reflect: true }) active = false;
+  @property({ reflect: true }) active = false;
 
-  @Watch("active")
-  activeHandler(): void {
-    if (this.disabled) {
-      this.active = false;
-    }
-  }
+  /** Specifies the appearance style of the component. */
+  @property({ reflect: true }) appearance: Extract<
+    "outline" | "outline-fill" | "solid" | "transparent",
+    Appearance
+  > = "solid";
+
+  /** When `true`, interaction is prevented and the component is displayed with lower opacity. */
+  @property({ reflect: true }) disabled = false;
 
   /** Specifies the icon used for the dropdown menu. */
-  @Prop({ reflect: true }) dropdownIconType: DropdownIconType = "chevron";
+  @property({ reflect: true }) dropdownIconType: DropdownIconType = "chevron";
 
   /** Accessible name for the dropdown menu. */
-  @Prop({ reflect: true }) dropdownLabel: string;
+  @property({ reflect: true }) dropdownLabel: string;
 
-  /**
-   * Specifies the component's fallback slotted content `placement` when it's initial or specified `placement` has insufficient space available.
-   */
-  @Prop() flipPlacements: FlipPlacement[];
+  /** Specifies the component's fallback slotted content `placement` when it's initial or specified `placement` has insufficient space available. */
+  @property() flipPlacements: FlipPlacement[];
 
-  /**
-    When `true`, a busy indicator is displayed on the primary button.
-   */
-  @Prop({ reflect: true }) loading = false;
+  /** Specifies the kind of the component, which will apply to border and background, if applicable. */
+  @property({ reflect: true }) kind: Extract<"brand" | "danger" | "inverse" | "neutral", Kind> =
+    "brand";
+
+  /** When `true`, a busy indicator is displayed on the primary button. */
+  @property({ reflect: true }) loading = false;
 
   /**
    * Determines the type of positioning to use for the overlaid content.
@@ -99,90 +89,108 @@ export class SplitButton implements InteractiveComponent, LoadableComponent {
    * Using `"absolute"` will work for most cases. The component will be positioned inside of overflowing parent containers and will affect the container's layout.
    *
    * `"fixed"` should be used to escape an overflowing parent container, or when the reference element's `position` CSS property is `"fixed"`.
-   *
    */
-  @Prop({ reflect: true }) overlayPositioning: OverlayPositioning = "absolute";
+  @property({ reflect: true }) overlayPositioning: OverlayPositioning = "absolute";
 
   /**
    * Determines where the component will be positioned relative to the container element.
    *
    * @default "bottom-end"
    */
-  @Prop({ reflect: true }) placement: MenuPlacement = "bottom-end";
+  @property({ reflect: true }) placement: MenuPlacement = "bottom-end";
 
   /** Specifies an icon to display at the end of the primary button. */
-  @Prop({ reflect: true }) primaryIconEnd: IconNameOrString;
+  @property({ reflect: true }) primaryIconEnd: IconNameOrString;
 
-  /**  Displays the `primaryIconStart` and/or `primaryIconEnd` as flipped when the element direction is right-to-left (`"rtl"`). */
-  @Prop({ reflect: true }) primaryIconFlipRtl: FlipContext;
+  /** Displays the `primaryIconStart` and/or `primaryIconEnd` as flipped when the element direction is right-to-left (`"rtl"`). */
+  @property({ reflect: true }) primaryIconFlipRtl: FlipContext;
 
   /** Specifies an icon to display at the start of the primary button. */
-  @Prop({ reflect: true }) primaryIconStart: IconNameOrString;
+  @property({ reflect: true }) primaryIconStart: IconNameOrString;
 
   /** Accessible name for the primary button. */
-  @Prop({ reflect: true }) primaryLabel: string;
+  @property({ reflect: true }) primaryLabel: string;
 
   /** Text displayed in the primary button. */
-  @Prop({ reflect: true }) primaryText: string;
+  @property({ reflect: true }) primaryText: string;
 
   /** Specifies the size of the component. */
-  @Prop({ reflect: true }) scale: Scale = "m";
+  @property({ reflect: true }) scale: Scale = "m";
 
   /** Specifies the width of the component. */
-  @Prop({ reflect: true }) width: Width = "auto";
+  @property({ reflect: true }) width: Width = "auto";
 
-  /**
-   * Fires when the primary button is clicked.
-   */
-  @Event({ cancelable: false })
-  calciteSplitButtonPrimaryClick: EventEmitter<void>;
+  // #endregion
 
-  /**
-   * Fires when the dropdown menu is clicked.
-   */
-  @Event({ cancelable: false })
-  calciteSplitButtonSecondaryClick: EventEmitter<void>;
-
-  //--------------------------------------------------------------------------
-  //
-  //  Public Methods
-  //
-  //--------------------------------------------------------------------------
+  // #region Public Methods
 
   /** Sets focus on the component's first focusable element. */
-  @Method()
+  @method()
   async setFocus(): Promise<void> {
     await componentFocusable(this);
-    this.el.focus();
+    focusFirstTabbable(this.el);
   }
 
-  //--------------------------------------------------------------------------
-  //
-  //  Private Properties
-  //
-  //--------------------------------------------------------------------------
+  // #endregion
 
-  @Element() el: HTMLCalciteSplitButtonElement;
+  // #region Events
 
-  //--------------------------------------------------------------------------
-  //
-  //  Lifecycle
-  //
-  //--------------------------------------------------------------------------
+  /** Fires when the primary button is clicked. */
+  calciteSplitButtonPrimaryClick = createEvent({ cancelable: false });
 
-  componentWillLoad(): void {
+  /** Fires when the dropdown menu is clicked. */
+  calciteSplitButtonSecondaryClick = createEvent({ cancelable: false });
+
+  // #endregion
+
+  // #region Lifecycle
+
+  load(): void {
     setUpLoadableComponent(this);
   }
 
-  componentDidLoad(): void {
-    setComponentLoaded(this);
+  override willUpdate(changes: PropertyValues<this>): void {
+    /* TODO: [MIGRATION] First time Lit calls willUpdate(), changes will include not just properties provided by the user, but also any default values your component set.
+    To account for this semantics change, the checks for (this.hasUpdated || value != defaultValue) was added in this method
+    Please refactor your code to reduce the need for this check.
+    Docs: https://qawebgis.esri.com/arcgis-components/?path=/docs/lumina-transition-from-stencil--docs#watching-for-property-changes */
+    if (changes.has("disabled") && (this.hasUpdated || this.disabled !== false)) {
+      if (!this.disabled) {
+        this.active = false;
+      }
+    }
+
+    if (changes.has("active") && (this.hasUpdated || this.active !== false)) {
+      if (this.disabled) {
+        this.active = false;
+      }
+    }
   }
 
-  componentDidRender(): void {
+  override updated(): void {
     updateHostInteraction(this);
   }
 
-  render(): VNode {
+  loaded(): void {
+    setComponentLoaded(this);
+  }
+
+  // #endregion
+
+  // #region Private Methods
+  private calciteSplitButtonPrimaryClickHandler(): void {
+    this.calciteSplitButtonPrimaryClick.emit();
+  }
+
+  private calciteSplitButtonSecondaryClickHandler(): void {
+    this.calciteSplitButtonSecondaryClick.emit();
+  }
+
+  // #endregion
+
+  // #region Rendering
+
+  override render(): JsxNode {
     const buttonWidth = this.width === "auto" ? "auto" : "full";
 
     return (
@@ -191,9 +199,9 @@ export class SplitButton implements InteractiveComponent, LoadableComponent {
           <calcite-button
             appearance={this.appearance}
             disabled={this.disabled}
-            icon-end={this.primaryIconEnd ? this.primaryIconEnd : null}
-            icon-start={this.primaryIconStart ? this.primaryIconStart : null}
+            iconEnd={this.primaryIconEnd ? this.primaryIconEnd : null}
             iconFlipRtl={this.primaryIconFlipRtl ? this.primaryIconFlipRtl : null}
+            iconStart={this.primaryIconStart ? this.primaryIconStart : null}
             kind={this.kind}
             label={this.primaryLabel}
             loading={this.loading}
@@ -216,12 +224,12 @@ export class SplitButton implements InteractiveComponent, LoadableComponent {
             overlayPositioning={this.overlayPositioning}
             placement={this.placement}
             scale={this.scale}
-            width-scale={this.scale}
+            widthScale={this.scale}
           >
             <calcite-button
               appearance={this.appearance}
               disabled={this.disabled}
-              icon-start={this.dropdownIcon}
+              iconStart={this.dropdownIcon}
               kind={this.kind}
               label={this.dropdownLabel}
               scale={this.scale}
@@ -236,19 +244,5 @@ export class SplitButton implements InteractiveComponent, LoadableComponent {
     );
   }
 
-  private calciteSplitButtonPrimaryClickHandler = (): CustomEvent =>
-    this.calciteSplitButtonPrimaryClick.emit();
-
-  private calciteSplitButtonSecondaryClickHandler = (): CustomEvent =>
-    this.calciteSplitButtonSecondaryClick.emit();
-
-  private get dropdownIcon(): string {
-    return this.dropdownIconType === "chevron"
-      ? "chevronDown"
-      : this.dropdownIconType === "caret"
-        ? "caretDown"
-        : this.dropdownIconType === "ellipsis"
-          ? "ellipsis"
-          : "handle-vertical";
-  }
+  // #endregion
 }